Refactor types

+++ native/quarto_ai/src/lib.rs
@@ -5,7 +5,7 @@ use std::{thread, time};
 
 // rustler does not support generics currently
 #[rustler::nif]
-fn choose_position_and_next_piece(board: Term, active_piece: i32) -> (usize, usize) {
+fn choose_position_and_next_piece(board: Term, active_piece: i32) -> (usize, i32) {
     let positions = get_tuple(board).expect("Error getting board tuple.");
     let mut empty_positions = Vec::new();
     let mut current_pieces = HashSet::from([active_piece]);
@@ -29,7 +29,7 @@ fn choose_position_and_next_piece(board: Term, active_piece: i32) -> (usize, usi
     let all_pieces: HashSet<i32> = (0..16).collect();
     let remaining_pieces: Vec<&i32> = all_pieces.difference(&current_pieces).collect();
     let random_piece_idx: usize = rng.gen_range(0..remaining_pieces.len());
-    let chosen_piece = *remaining_pieces[random_piece_idx] as usize;
+    let chosen_piece = *remaining_pieces[random_piece_idx];
 
     // sleep 1 second for random AI
     let one_second = time::Duration::from_secs(1);
